ProyectoFinal/requirements.txt
altgraph==0.17.4
asgiref==3.7.2
attrs==23.1.0
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
colorama==0.4.6
cryptography==42.0.7
distlib==0.3.8
dj-database-url==2.1.0
Django==5.0
django-allauth==0.62.1
django-shell-plus==1.1.7
Eel==0.16.0
et-xmlfile==1.1.0
filelock==3.13.3
future==0.18.3
gevent==23.9.1
gevent-websocket==0.10.1
greenlet==3.0.1
h11==0.14.0
idna==3.4
iniconfig==2.0.0
jsonpatch==1.33
jsonpointer==2.4
jwt==1.3.1
MouseInfo==0.1.3
networkx==3.1
numpy==1.26.0
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.1
pefile==2023.2.7
Pillow==10.0.1
platformdirs==4.2.0
pluggy==1.5.0
plus==1994.12.4
plyer==2.1.0
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.11
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pytest==8.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytweening==1.0.7
pytz==2023.3.post1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
Random-Word==1.0.11
requests==2.31.0
scipy==1.11.3
selenium==4.17.2
shell==1.0.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tornado==6.3.3
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.6
virtualenv==20.25.1
visdom==0.2.4
websocket-client==1.6.4
whichcraft==0.6.1
wsproto==1.2.0
zope.event==5.0
zope.interface==6.1
